python/requirements.txt
boto3==1.11.9
PyYAML==5.3
